+++ src/interpreter/function.h
@@ -6,7 +6,7 @@
 namespace ts {
 	class Function : public InstructionContainer {
 		public:
-			Function(Instruction* head, size_t argumentCount, size_t variableCount, string functionName, string namespaceName = string());
+			Function(Instruction* head, uint64_t argumentCount, uint64_t variableCount, string functionName, string namespaceName = string());
 			Function(sl::Function* function);
 			~Function();
 
@@ -14,8 +14,8 @@ namespace ts {
 			sl::Function* function = nullptr; // the standard library function to call if we're a standard library function
 			string functionName;
 			string namespaceName;
-			size_t argumentCount;
-			size_t variableCount;
+			uint64_t argumentCount;
+			uint64_t variableCount;
 			bool isActive = true;
 			bool isPackaged;
 	};
